refactor(gradle): move replacingSerializer to test/util.ts
